Mesh Chunking (#96)
Updates the way we represent meshes internally. Prior to this, we would treat all geometry data as a singular "mesh". This was convenient as it meant we could do a single draw call to render the geometry. However, this makes it difficult to support things like meshes that support multiple materials which also requires multiple draw calls.

In this update, we add support for both multiple meshes as well as multiple mesh "chunks" within each mesh. Where a chunk is meant to represent a singular renderable set of geometry with a single material.